Move search key to QuickSearch registration
Rather than have a source know about what should be done when a selection is made (which is what the filterKey was saying), this pushes that logic to the registration of the controller.

In this case, the search key is simply the parameter name for the filter. However, the search key could be anything: a class name, a method name, or anything that enables the listener to the selection event to determine what do to.