Add Quick Search to Locations
This called out that the Jump To source, if it includes links in the items, must be a separate source than a non-Jump To. If the redirect to the target were to be deferred to the handler, a single source could be used (though it would still need to be able to adjust the template header). It seems simpler to have two sources, which can share logic through inheritance.