Add default styles to elements
This adds default styles to elements rather than requiring them to use a class by default.

The one hedge here is the class on the form. That might make it easy to have one set of styles for self service pages and another for admin pages. Even then, one could be the default.