Switch to Bulma Phlex (#65)
* Remove Tailwind

This removes the tailwindcss-rails gem, along with some of the related artifacts. For example, since the tailwind watch process is no longer required, the Procfile.dev has been removed and the bin/dev has been simplified.

* Add phlex-rails

* Add layout for Phlex views

* Switch from sprockets to propshaft

* Update JS import statements

Do not use relative imports.

* Add Phlexable mixin to ApplicationController

This makes it possible for controllers to leverage conventions:

  render phlex(@films)

* Add literal gem

* Update integrity sha on Font Awesome stylesheet

* Show Film and Poster

This updates the Show actions for Film and Poster.

* Add openssl to Gemfile

There's an open issue that can be worked around by explicitly adding the openssl gem.

See https://github.com/ruby/openssl/issues/949

* Update gems

Pin minitest at 5.

* Films Index, Quick Search

This updates the films index, retiring most of the custom CSS in favor of Bulma styles/classes.

* Companies: Index and Show

* Add poster to Films index eager loading

* Remove redundant DOMID helper

It is in the Components::Base class.

* Introduce FilmCredit mixin

This handles the two belongs to associations (film and person) and adds a new delegation to provide `film_name`.

* People: Index and Show

* Rubocop corrections

* Removed unused Film Locations views

There's no controller nor routes for these.

* Extract QuickSearch component

* Value object: Search Parameters

Encapsulates the logic to parse and build the search parameters from the request params. This simplifies what gets passed to the views and then onto the QuickSearch component.

Also...
- the QuickSearch property `sorts` was renamed to `sort_options`
- in the Films index view, the arguments for `search_sources` and `sort_options` were broken out into methods for readability

* Locations: Index and Show

* Remove Pagination

The helper and views have been replaced by Bulma Phlex.

* Add dummy ApplicationMailer